Support go test -count=X: only call sql.Register once
It is useful to run go test with -count=X to check for flaky tests.
Without this change, this fails because calling sql.Register with the
same name more than once panics. This introduces two new functions,
one that only registers a custom driver once, and one that registers a
custom driver every time. This allows the tests to run in a loop.